Refactor git client functionality into GitClientHandler
