+++ src/main/java/org/jabref/logic/git/GitHandler.java
@@ -7,8 +7,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Optional;
 
-import org.jabref.logic.preferences.AutoPushMode;
-
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.RmCommand;
 import org.eclipse.jgit.api.Status;
@@ -24,7 +22,6 @@ import org.slf4j.LoggerFactory;
  * This class handles the updating of the local and remote git repository that is located at the repository path
  * This provides an easy-to-use interface to manage a git repository
  */
-@SuppressWarnings("checkstyle:RegexpMultiline")
 public class GitHandler {
     static final Logger LOGGER = LoggerFactory.getLogger(GitHandler.class);
     final Path repositoryPath;
@@ -195,7 +192,7 @@ public class GitHandler {
                    .setCredentialsProvider(credentialsProvider)
                    .call();
             } catch (GitAPIException e) {
-                LOGGER.info("Failed to push: ".concat(e.toString()));
+                LOGGER.info("Failed to push: ");
             }
         }
     }
@@ -207,7 +204,7 @@ public class GitHandler {
                    .setCredentialsProvider(credentialsProvider)
                    .call();
             } catch (GitAPIException e) {
-                LOGGER.info("Failed to pull".concat(e.toString()));
+                LOGGER.info("Failed to pull");
             }
         }
     }
@@ -216,32 +213,5 @@ public class GitHandler {
         try (Git git = Git.open(this.repositoryPathAsFile)) {
             return git.getRepository().getBranch();
         }
-    }
-
-    /**
-     * Contains logic for commiting and pushing after a database is saved locally,
-     * if the relevant preferences are present.
-     *
-     * @param preferences preferences for git
-     */
-    public void postSaveDatabaseAction(GitPreferences preferences) {
-        if (this.isGitRepository() &&
-                preferences.getAutoPushMode() == AutoPushMode.ON_SAVE &&
-                preferences.getAutoPushEnabled()) {
-            this.updateCredentials(preferences);
-            try {
-                this.createCommitOnCurrentBranch("Automatic update via JabRef", false);
-                this.pushCommitsToRemoteRepository();
-            } catch (GitAPIException | IOException e) {
-                LOGGER.info("Failed to push".concat(e.toString()));
-            }
-        }
-    }
-
-    public void updateCredentials(GitPreferences preferences) {
-        this.credentialsProvider = new UsernamePasswordCredentialsProvider(
-                preferences.getGitHubUsername(),
-                preferences.getGitHubPasskey()
-        );
     }
 }
